+++ src/dstack/_internal/server/background/scheduled_tasks/instances.py
@@ -59,7 +59,7 @@ from dstack._internal.core.models.runs import (
     JobProvisioningData,
 )
 from dstack._internal.server import settings as server_settings
-from dstack._internal.server.background.tasks.common import get_provisioning_timeout
+from dstack._internal.server.background.scheduled_tasks.common import get_provisioning_timeout
 from dstack._internal.server.db import get_db, get_session_ctx
 from dstack._internal.server.models import (
     FleetModel,
