+++ src/dstack/_internal/server/background/scheduled_tasks/placement_groups.py
