+++ src/dstack/_internal/server/background/scheduled_tasks/submitted_jobs.py
@@ -57,7 +57,9 @@ from dstack._internal.core.models.runs import (
 from dstack._internal.core.models.volumes import Volume
 from dstack._internal.core.services.profiles import get_termination
 from dstack._internal.server import settings
-from dstack._internal.server.background.tasks.process_compute_groups import ComputeGroupStatus
+from dstack._internal.server.background.scheduled_tasks.compute_groups import (
+    ComputeGroupStatus,
+)
 from dstack._internal.server.db import (
     get_db,
     get_session_ctx,
