+++ src/tests/_internal/server/background/scheduled_tasks/test_process_events.py
@@ -6,7 +6,7 @@ from freezegun import freeze_time
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from dstack._internal.server import settings
-from dstack._internal.server.background.tasks.process_events import delete_events
+from dstack._internal.server.background.scheduled_tasks.events import delete_events
 from dstack._internal.server.services import events
 from dstack._internal.server.testing.common import create_user, list_events
